fix: 피드백 반영 수정사항
