+++ src/components/common/AttendanceItem/AttendanceItem.tsx
@@ -1,0 +1,11 @@
+const AttendanceItem = () => {
+  return (
+    <div className='flex flex-row justify-around gap-1 bg-white p-3 rounded-xl border-2 border-solid border-pink-500'>
+      <div className=''>누구누구 님 외 2명</div>
+      <div>000-0000-0000</div>
+      <div>O</div>
+      <div>O</div>
+    </div>
+  )
+}
+export default AttendanceItem
