+++ src/components/common/RsvpItem/RsvpItem.tsx
@@ -1,7 +1,3 @@
-
-
-
-
 interface RsvpDetail {
   title: string;
   description?: string;
@@ -13,8 +9,6 @@ interface RsvpDetail {
 
 
 export default function RsvpItem({ total, title, description, attend, unattend, bride }: RsvpDetail) {
-
-
   return (
     <div className="flex flex-row justify-between text-gray-700 items-center p-2 shadow-md border border-gray-100 rounded-lg bg-white">
       <div className={`flex flex-col items-start ${total && 'font-semibold p-3 text-2xl'}`}>
