+++ src/components/display/PhotoTalkSection/PhotoTalk.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 const PhotoTalkSection = () => {
   return <div className="column-center">RealTimePhotoSection</div>;
 };
